Moved FunSpec tests into its own module.
