+++ build.sbt
@@ -19,6 +19,7 @@ lazy val scalatestDiagramsTest = ScalatestBuild.scalatestDiagramsTest
 lazy val scalatestFeatureSpecTest = ScalatestBuild.scalatestFeatureSpecTest
 lazy val scalatestFlatSpecTest = ScalatestBuild.scalatestFlatSpecTest
 lazy val scalatestFreeSpecTest = ScalatestBuild.scalatestFreeSpecTest
+lazy val scalatestFunSpecTest = ScalatestBuild.scalatestFunSpecTest
 lazy val scalatestJS = ScalatestBuild.scalatestJS
 lazy val scalatestNative = ScalatestBuild.scalatestNative
 lazy val scalatestDotty = ScalatestBuild.scalatestDotty
@@ -27,16 +28,19 @@ lazy val scalatestDiagramsTestJS = ScalatestBuild.scalatestDiagramsTestJS
 lazy val scalatestFeatureSpecTestJS = ScalatestBuild.scalatestFeatureSpecTestJS
 lazy val scalatestFlatSpecTestJS = ScalatestBuild.scalatestFlatSpecTestJS
 lazy val scalatestFreeSpecTestJS = ScalatestBuild.scalatestFreeSpecTestJS
+lazy val scalatestFunSpecTestJS = ScalatestBuild.scalatestFunSpecTestJS
 lazy val scalatestTestNative = ScalatestBuild.scalatestTestNative
 lazy val scalatestDiagramsTestNative = ScalatestBuild.scalatestDiagramsTestNative
 lazy val scalatestFeatureSpecTestNative = ScalatestBuild.scalatestFeatureSpecTestNative
 lazy val scalatestFlatSpecTestNative = ScalatestBuild.scalatestFlatSpecTestNative
 lazy val scalatestFreeSpecTestNative = ScalatestBuild.scalatestFreeSpecTestNative
+lazy val scalatestFunSpecTestNative = ScalatestBuild.scalatestFunSpecTestNative
 lazy val scalatestTestDotty = ScalatestBuild.scalatestTestDotty
 lazy val scalatestDiagramsTestDotty = ScalatestBuild.scalatestDiagramsTestDotty
 lazy val scalatestFeatureSpecTestDotty = ScalatestBuild.scalatestFeatureSpecTestDotty
 lazy val scalatestFlatSpecTestDotty = ScalatestBuild.scalatestFlatSpecTestDotty
 lazy val scalatestFreeSpecTestDotty = ScalatestBuild.scalatestFreeSpecTestDotty
+lazy val scalatestFunSpecTestDotty = ScalatestBuild.scalatestFunSpecTestDotty
 lazy val scalatestApp = ScalatestBuild.scalatestApp
 lazy val scalatestAppJS = ScalatestBuild.scalatestAppJS
 lazy val scalatestAppNative = ScalatestBuild.scalatestAppNative
