+++ jvm/funspec-test/src/test/scala/org/scalatest/funspec/AsyncFunSpecLikeSpec.scala
@@ -13,9 +13,13 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.scalatest
+package org.scalatest.funspec
 
-import SharedHelpers.EventRecordingReporter
+import org.scalatest.SharedHelpers.EventRecordingReporter
+import org.scalatest.funspec
+import org.scalatest.ParallelTestExecution
+import org.scalatest.Args
+import org.scalatest.Assertion
 import scala.concurrent.{Promise, ExecutionContext, Future}
 import org.scalatest.concurrent.SleepHelper
 import org.scalatest.events.{InfoProvided, MarkupProvided}
@@ -27,6 +31,7 @@ class AsyncFunSpecLikeSpec extends funspec.AnyFunSpec {
 
   describe("AsyncFunSpecLike") {
 
+    // SKIP-DOTTY-START
     it("can be used for tests that return Future under parallel async test execution") {
 
       class ExampleSpec extends AsyncFunSpecLike with ParallelTestExecution {
@@ -56,13 +61,13 @@ class AsyncFunSpecLikeSpec extends funspec.AnyFunSpec {
 
         it("test 4") {
           Future {
-            cancel
+            cancel()
           }
         }
 
         ignore("test 5") {
           Future {
-            cancel
+            cancel()
           }
         }
 
@@ -110,11 +115,11 @@ class AsyncFunSpecLikeSpec extends funspec.AnyFunSpec {
         }
 
         it("test 4") {
-          cancel
+          cancel()
         }
 
         ignore("test 5") {
-          cancel
+          cancel()
         }
 
         override def newInstance = new ExampleSpec
@@ -138,6 +143,7 @@ class AsyncFunSpecLikeSpec extends funspec.AnyFunSpec {
       assert(rep.testIgnoredEventsReceived.length == 1)
       assert(rep.testIgnoredEventsReceived(0).testName == "test 5")
     }
+    // SKIP-DOTTY-END
 
     it("should run tests that return Future in serial by default") {
