+++ jvm/funspec-test/src/test/scala/org/scalatest/funspec/AsyncFunSpecSpec.scala
@@ -13,9 +13,15 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.scalatest
+package org.scalatest.funspec
 
-import SharedHelpers.{EventRecordingReporter, thisLineNumber}
+import org.scalatest.SharedHelpers.{EventRecordingReporter, thisLineNumber}
+import org.scalatest.funspec
+import org.scalatest.Args
+import org.scalatest.ParallelTestExecution
+import org.scalatest.Assertion
+import org.scalatest.FailureMessages
+import org.scalatest.UnquotedString
 import scala.concurrent.{Promise, ExecutionContext, Future}
 import org.scalatest.concurrent.SleepHelper
 import org.scalatest.events.{InfoProvided, MarkupProvided}
@@ -31,6 +37,7 @@ class AsyncFunSpecSpec extends funspec.AnyFunSpec {
 
   describe("AsyncFunSpec") {
 
+    // SKIP-DOTTY-START
     it("can be used for tests that return a Future under parallel async test execution") {
 
       class ExampleSpec extends AsyncFunSpec with ParallelTestExecution /* with expectations.Expectations */ { // Can resurrect expectations.Expectations tests later
@@ -60,13 +67,13 @@ class AsyncFunSpecSpec extends funspec.AnyFunSpec {
 
         it("test 4") {
           Future {
-            cancel
+            cancel()
           }
         }
 
         ignore("test 5") {
           Future {
-            cancel
+            cancel()
           }
         }
 
@@ -133,11 +140,11 @@ class AsyncFunSpecSpec extends funspec.AnyFunSpec {
         }
 
         it("test 4") {
-          cancel
+          cancel()
         }
 
         ignore("test 5") {
-          cancel
+          cancel()
         }
 
         it("test 6") {
@@ -177,6 +184,7 @@ class AsyncFunSpecSpec extends funspec.AnyFunSpec {
       assert(rep.testIgnoredEventsReceived.length == 1)
       assert(rep.testIgnoredEventsReceived(0).testName == "test 5")
     }
+    // SKIP-DOTTY-END
 
     it("should run tests that return Future in serial by default") {
